Remove non-available builds from filesystem.
When the language isn't supported by JoliCi, the build directories were
still being left around.  These build directories would then cause
errors when the toplevel directory from the Run command would fail on
rmdir() (RunCommand.php:75).

It still would be better if a warning could be emitted back to the user,
but that would require some extensive changes in order to pass an error
back out of the build strategy class or restructure things so that the
RunCommand can be aware that there were some builds that couldn't be
created.